feat(ui,site): upgrade, update /about
