+++ packages/ui/src/components/primitives/ui/button.tsx
@@ -11,6 +11,7 @@ const buttonVariants = cva(
 		variants: {
 			variant: {
 				default: 'bg-primary text-primary-foreground hover:bg-primary/90',
+				blue: 'bg-ao-blue text-primary-foreground dark:text-primary hover:bg-ao-blue/90',
 				destructive:
 					'bg-destructive text-destructive-foreground hover:bg-destructive/90',
 				outline:
